Unify process into a single file
